[gardening] Sink compiler invocation code into libIDE
